// src/types.ts
export interface PromptItem {
  id: string;
  name: string;
  prompt: string;
}

export interface PromptFolder {
  id: string;
  name: string;
  presets: PromptItem[];
  // FIX: Add optional showInQuickBar property to fix type errors
  showInQuickBar?: boolean;
}